commiting some refactoring of the engine to switch to a dedicated branch

+++ tinyutils.h
@@ -1,11 +1,9 @@
-
-
 #ifndef TINY_UTILS
 #define TINY_UTILS
 
 typedef struct
 {
-    int x, y;
+    float x, y;
 }vec2_t;
 
 // void draw_pixel(int x, int y, color_t color, GameRenderer game_renderer);
